Bluetooth: Unregister callbacks when turning off
In current implementation, callbacks registered for notifying
Bluetooth state change is not unregistered when turning off
Bluetooth. This can cause the binder leakage in Bluetooth
On/Off stress testing. Number of receiver is increased after
each turn of Bluetooth On/Off.
The solution is to unregister callbacks after calling doUnbind().

Change-Id: Id6628db4dac841db20a49522b4ddf5db0c85d482