wifi: Remove @SystemApi from value-added getBands() API
The API with @SystemApi annotation means it is already upstreamed.
This is validated by CtsSystemApiAnnotationTestCases:
android.signature.cts.api.AnnotationTest#testAnnotation

For value-added getBands() API, it is marked as @SystemApi but not
upstreamed to android11-tests-dev, so CTS failed with message:
"... SoftApConfiguration.getBands() Error: Method annotated with
@android.annotation.SystemApi(client=PRIVILEGED_APPS) does not
exist in the API."

This CL removes @SystemApi from value-added getBands() API, and
updates system-current.txt.

CRs-Fixed: 3014698
Change-Id: I50f78fd92403eb8bd076662a2e50a7b97b4df6ba