Fix behavior of Notification shade when manually closed without opening it completely
Issue: When open and close the notification shade immediatly with one
swipe, the widgets like Maps or Bluetooth are not clickable. Instead
when opening the notification shade from the top with a swipe the shade
is shown over the whole area immediately without animation. This is also
the reason why taps are not received by Maps, because they are consumed
by the shade touch listener.

Fix: Handle the else case of the open touch listener to allow closing
the shade correctly.

This is a modified cherry pick of Change-Id: Ib124ea1df99f6f70ebd74e62fc41c477e47e14f0
as the names of the files were changed.

Bug: b/198749961
Test: Manual UI testing
Change-Id: I072f942f2345d473a37fb28cde3585401eb962c3