WiredAccessoryManager: support for DP/HDMI display on soc:qcom,msm-ext-disp.
Issue : The DP was not getting detected when connected kona HDK to external display.

Root Cause : As per the design the controller value should be 0 for DP and 1 for HDMI,
The correct controller value was not selected when DP was connected due to which
the routing did not happen to external display.

Fix : Create API setControllerIdx to set the controller index based on the interface
which was selected.

Change-Id: Id5987315623d4eb2ec8b9369efff3645cda4b677