WiredAccessoryManager: support HDMI detection when already connected before boot
Issue : HDMI was not detected if connected to device before boot.

Root Cause : When device boots up the function onUEvent is not called and init
is called. onUEvent is called after boot up during Hotplug. There is no mechanism
in init function to update the interface when the sysfs node state is read as 1.

Fix : Added the mechanism to update the interface to HDMI or DP based on the state
read from the sysfs node accordingly.

Change-Id: I28cd4433348491f71c06415ca8f42869d38ff82a