Dismiss keyguard when simpin auth'd and...
security method is none. This is mostly to fix the case where we auth
sim pin in the set up wizard and it goes straight to keyguard instead of
the setup wizard activity.

This works with the prevent bypass keyguard flag because the device
should be noe secure in this case.

Fixes: 222446076
Test: turn locked sim on, which opens the sim pin screen. Auth the
screen and observe that keyguard is not shown.

(cherry picked from commit 48fa9bef3451e4a358c941af5b230f99881c5cb6)
Change-Id: Ib2d5400e930fa4d6f7f98e46f32a290ba5f558f4