perf: Added changes for perf-lock APIs calls from activity launcher and swipe gestures
Changes are added in the Activity Starter files for getting the cpu frequencies boosted
when the application is launched. Similar changes are added to the swipe gesture files
for boosting the frequencies when swiping action is performed. Zygote change is for
loading the jar file for the perf-core service.

Change-Id: I81e7b48891e6c44cc28a792b4533b2143040f934
Signed-off-by: Ankit Kumar Panda <quic_ankipand@quicinc.com>
(cherry picked from commit 7c6410eb3b19738c51e3974197a941205438c0f0)