Fixes update ownership preapproval flow
This change sets the package name when computing commit preapproval
reason. Previously, with the package name unavailable that path would
not be able to look up the existing package and so wouldn't know the
owner. Additionally, if the installer required user action, the wrong
dialog would have been displayed when the update ownership dialog should
have been. This fixes both issues.

Add new change id to check whether the build includes the fix or not.
If the change is disabled, it means the build includes the fix. If
the build doesn't include the change id, when the caller checks the
change, it is enabled.

Bug: 293644536
Test: atest PreapprovalInstallTest UpdateOwnershipEnforcementTest
Change-Id: I210fe1bb13a2096eb7e087806298ed1d4fa4905d
Merged-In: I210fe1bb13a2096eb7e087806298ed1d4fa4905d
 (cherry picked from commit 2eb995c054335b85c6ef72cc302500d14042f2d9)