RESTRICT AUTOMERGE Log to detect usage of whitelistToken when sending non-PI target
Log ActivityManagerService.sendIntentSender if the target is not a
PendingIntent and a non-null whitelistToken is sent to the client.
This is simply to detect if there are real cases this would happen
before we decide simply remove whitelistToken in that case.

Do not pass  whitelistToken when sending non-PI target

In ActivityManagerService.sendIntentSender, if the target is not a
PendingIntent, do not send whitelistToken to the client.

Bug: 279428283
Test: Manual test
Change-Id: I017486354a1ab2f14d0472c355583d53c27c4810
(cherry picked from commit 7a76717b61d8cb90a4987454f34e88417d68608b)