Set no data transfer on function switch timeout for accessory mode
In case of function switch times out, we will check whether
the last function set was accessory. If this is the case, it is
recommended to set the function to NONE(No data transfer) rather than
setting it to the default USB function.

Bug: 353712853
Test: Build the code, flash the device and test it.
Test: atest CtsUsbManagerTestCases
Test: run CtsVerifier tool
Test: atest CtsUsbTests
Change-Id: I698e9df0333cbb51dd9bd5917a94d81273a2784a
(cherry picked from commit 7c6ec68537ba8abf798afd9ab7c3e5889841171f)